more descriptive log messages & change folder name
