+++ src/events/guild/scheduledEvents/guildScheduledEventDelete.ts
